requirements.txt
Flask
Flask-SQLAlchemy
Flask-RESTful
pywebview
pyinstaller
